fix : sua chuyen phong va import

+++ src/main/java/iuh/fit/se/group1/dto/TransferResult.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class TransferResult implements java.io.Serializable {
     private boolean success;
     private String message;
-    private long surcharge;
+    private double surcharge;
 }
